* Fix improper signed int conversion of IMAP uid and msn values.
 * Change imap literal counts to parse and store unsigned ints.
 * Fix imap status count range check.
 * cmd_handle_fatal: make error message a bit more descriptive
 * Updated French translation.
 * Create pgp and s/mime default and sign_as key vars. (see #3983)
 * Add missing setup calls when resuming encrypted drafts.
 * mutt_pretty_size: show real number for small files
 * examine_directory: set directory/symlink size to zero
 * Update pl.po
 * Fixed GPGME translations that weren’t shown but affected the keyboard
 * docs: update encrypt-to-self
 * Update smime.rc: Typo fix, consistent headings
 * Add pgp_default_key and smime_sign_as info to contrib rc files.
 * Split Copyright and Thanks in help output.
 * strip out copyright translations

+++ imap/imap_private.h
@@ -279,7 +279,7 @@ int imap_exec_msgset(struct ImapData *idata, const char *pre, const char *post,
 int imap_open_connection(struct ImapData *idata);
 void imap_close_connection(struct ImapData *idata);
 struct ImapData *imap_conn_find(const struct Account *account, int flags);
-int imap_read_literal(FILE *fp, struct ImapData *idata, long bytes, struct Progress *pbar);
+int imap_read_literal(FILE *fp, struct ImapData *idata, unsigned long bytes, struct Progress *pbar);
 void imap_expunge_mailbox(struct ImapData *idata);
 void imap_logout(struct ImapData **idata);
 int imap_sync_message_for_copy(struct ImapData *idata, struct Header *hdr, struct Buffer *cmd, int *err_continue);
@@ -324,7 +324,7 @@ struct ImapData *imap_new_idata(void);
 void imap_free_idata(struct ImapData **idata);
 char *imap_fix_path(struct ImapData *idata, const char *mailbox, char *path, size_t plen);
 void imap_cachepath(struct ImapData *idata, const char *mailbox, char *dest, size_t dlen);
-int imap_get_literal_count(const char *buf, long *bytes);
+int imap_get_literal_count(const char *buf, unsigned int *bytes);
 char *imap_get_qualifier(char *buf);
 int imap_mxcmp(const char *mx1, const char *mx2);
 char *imap_next_word(char *s);
